+++ NetTeam4/NetTeam4/src/Game/Player.cpp
@@ -4,8 +4,8 @@
 void PlayerMessage::Read(BinaryStream* stream, NetworkManager& manager)
 {
 	Id = stream->Read<int>();
-	x = stream->Read<int>();
-	y = stream->Read<int>();
+	x = stream->Read<float>();
+	y = stream->Read<float>();
 	frameId = stream->Read<int>();
 
 	Client& client = (Client&)manager;
@@ -15,8 +15,8 @@ void PlayerMessage::Read(BinaryStream* stream, NetworkManager& manager)
 int PlayerMessage::Write(BinaryStream* stream){
 
 	stream->Write<int>(Id);
-	stream->Write<int>(x);
-	stream->Write<int>(y);
+	stream->Write<float>(x);
+	stream->Write<float>(y);
 	stream->Write<int>(frameId);
 
 	return sizeof(int) * 4;
